server/cloudcluster/requirements.txt
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-core==2.13.16
aliyun-python-sdk-ecs==4.19.4
aliyun-python-sdk-nas==3.9.0
aliyun-python-sdk-ram==3.2.0
aliyun-python-sdk-slb==3.2.17
ansible==2.9.16
azure-common==1.1.25
azure-graphrbac==0.61.1
azure-identity
azure-mgmt-authorization==0.61.0
azure-mgmt-compute==12.0.0
azure-mgmt-network==11.0.0
azure-mgmt-resource==11.0.0
azure-mgmt-storage
azure-mgmt-subscription==0.5.0
azure-storage-blob==12.12.0
boto3==1.13.5
celery==4.4.0
django-crypto-fields==0.1.26
django-extensions==2.2.8
django==3.0.3
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
edc-constants==0.2.28
edc-model-fields==0.1.11
google-api-python-client==2.54.0
google-auth-oauthlib==0.4.1
google-cloud-asset==3.7.1
google-cloud-storage==2.3.0
ipconflict==0.5.0
jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
mock==4.0.2
msal==1.2.0
mysqlclient==1.4.4
netaddr==0.7.19
openstacksdk==0.58.0
paramiko==2.6.0
pbr==5.4.4
ruamel.yaml==0.16.10
uWSGI==2.0.18
yamllint==1.20.0
requests==2.25.1
cryptography==2.8
Pillow==8.1.0
django_redis==4.12.1
django-prometheus==2.1.0
JSON-log-formatter==0.3.1
mysql-connector-python==8.0.26
MarkupSafe==2.0.1
python-wireguard==0.2.1
hvac==0.11.2
drf-yasg==1.20.0
importlib-metadata==4.13.0